+++ packages/textarea/src/mwc-textarea-base.ts
@@ -64,7 +64,7 @@ export abstract class TextAreaBase extends TextFieldBase {
           placeholder="${this.placeholder}"
           ?required="${this.required}"
           maxlength="${ifDefined(maxOrUndef)}"
-          @change="${this.handleInputChange}"
+          @input="${this.handleInputChange}"
           @blur="${this.onInputBlur}">
       </textarea>`;
   }